refactor: remote commands
