+++ internal/pkg/service/cli/cmd/remote/create/bucket/bucket.go
@@ -1,4 +1,4 @@
-package create
+package bucket
 
 import (
 	"time"
