+++ internal/pkg/service/cli/cmd/remote/create/table/table.go
@@ -1,4 +1,4 @@
-package create
+package table
 
 import (
 	"time"
